+++ ts/libs/polaris-nx/src/generators/predicted-metric-controller/generator.ts
@@ -1,5 +1,8 @@
+import * as child_process from 'child_process';
 import * as path from 'path';
-import { Generator, Tree, formatFiles, generateFiles, offsetFromRoot, readProjectConfiguration, updateProjectConfiguration } from '@nx/devkit';
+import { Generator, Tree, formatFiles, generateFiles, joinPathFragments, offsetFromRoot, readProjectConfiguration, updateProjectConfiguration } from '@nx/devkit';
+import { ObjectKind } from '@polaris-sloc/core';
+import { flushChanges } from 'nx/src/generators/tree';
 import {
     NPM_PACKAGES,
     POLARIS_INIT_LIB_FN_NAME,
@@ -13,6 +16,8 @@ import {
     createAppProject,
     getComposedMetricTypeNames,
     getContainerImageName,
+    getTempDir,
+    getWorkspaceTsConfigPath,
     normalizeProjectGeneratorOptions,
     runCallbacksSequentially,
 } from '../../util';
@@ -77,8 +82,12 @@ export default generateComposedMetricController;
 function addComposedMetricControllerFiles(host: Tree, options: NormalizedPredictedMetricControllerGeneratorSchema): void {
     const compMetricNames = getComposedMetricTypeNames(options.compMetricType);
 
+    const { group, kind } = extractCompMetricType(host, options, compMetricNames.compMetricType)
+
     const templateOptions = {
         ...compMetricNames,
+        compMetricApiGroup: group,
+        compMetricK8sResources: kind.toLocaleLowerCase(),
         compMetricTypePkg: options.compMetricTypePkg,
         initPolarisLibFn: POLARIS_INIT_LIB_FN_NAME,
         controllerProjectName: options.projectName,
@@ -89,4 +98,43 @@ function addComposedMetricControllerFiles(host: Tree, options: NormalizedPredict
 
     generateFiles(host, path.join(__dirname, 'files/composed-metric-controller'), options.projectRoot, templateOptions);
     generateTypeScriptDockerfile(host, options);
+}
+
+function extractCompMetricType(host: Tree, options: NormalizedPredictedMetricControllerGeneratorSchema, compMetricType: string): ObjectKind {
+    const tempDir = getTempDir(options.name, 'gen-composed-metric-controller');
+    generateAndWriteScripts(host, options.compMetricTypePkg, compMetricType, tempDir);
+
+    // Run the gen-crds script.
+    const scriptTsConfig = joinPathFragments(tempDir, 'tsconfig.json');
+    const scriptTs = joinPathFragments(tempDir, 'gen-controller.ts');
+    const result = child_process.spawnSync(
+        'npx',
+        [ 'ts-node', '--project', scriptTsConfig, scriptTs ],
+        {
+            cwd: host.root,
+        },
+    );
+
+    const objectKindString = result.stdout.toString();
+    return JSON.parse(objectKindString);
+}
+
+function generateAndWriteScripts(host: Tree, compMetricTypePkg: string, polarisType: string, tempDir: string): void {
+    const tsConfigBase = getWorkspaceTsConfigPath(host);
+    const templateOptions = {
+        tsConfigBase: joinPathFragments(offsetFromRoot(tempDir), tsConfigBase),
+        compMetricTypePkg,
+        polarisType,
+        template: '',
+    };
+
+    generateFiles(
+        host,
+        path.join(__dirname, 'files/gen-composed-metric-controller-scripts'),
+        tempDir,
+        templateOptions,
+    );
+
+    const changes = host.listChanges().filter((change) => change.path.startsWith(tempDir));
+    flushChanges(host.root, changes);
 }