+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 * Add `apply-slo-mapping` command to polaris-cli
 * Add `polaris-cli migrate` command that migrates all polaris packages to the specified version and all Nx packages to the appropriate version.
 * Use an executor for the `polaris-cli deploy` command and check if unmodifiable Kubernets objects already exist
-
+* Extract Group, Version & Kind from Polaris Types during Controller Generation
 
 
 ## v0.7.0 (2023-11-03)